feat: unify token cookie
