+++ CHANGELOG.md
@@ -318,3 +318,4 @@
 - CSRF-токен сохраняется в `localStorage` и передаётся в заголовке `X-XSRF-TOKEN`
 - При недоступности `localStorage` токен хранится в памяти, добавлен раздел в technical_manual.md
 - Страница входа показывает уведомления об ошибках авторизации
+- Функция `setTokenCookie` избавляет от дублирования кода установки cookie
