+++ README.md
@@ -139,6 +139,7 @@ curl -X POST "$APP_URL/api/v1/tasks" \
 - Исправлен путь к auth.service для стабильного запуска
 - Указано расширение `.ts` при импорте roles.decorator
 - Установка cookie `token` логируется с указанием домена
+- Добавлена функция `setTokenCookie` для единой установки cookie `token`
 
 ### Ротация токенов
