+++ ROADMAP.md
@@ -250,3 +250,4 @@
 220. Middleware `verifyToken` продлевает cookie `token` при каждом запросе
 221. CSRF-токен сохраняется в `localStorage` и отправляется в заголовке
 222. При недоступности `localStorage` токен хранится в памяти, страница входа показывает ошибки
+223. Функция `setTokenCookie` объединяет установку cookie `token`
