+++ docs/technical_manual.md
@@ -108,6 +108,7 @@ Middleware `checkRole` и `checkTaskAccess` записывают отказ до
    токен недействителен.
 4. Для защиты от XSS cookie `token` имеет флаги `HttpOnly` и `SameSite=Lax`. В продакшене добавляется `Secure` и используется домен из `COOKIE_DOMAIN` либо `APP_URL`. В режиме разработки cookie передаётся и по HTTP без явного домена. Токен живёт семь дней.
 5. Middleware `verifyToken` продлевает cookie `token` при каждом запросе,
+6. Отдельная функция `setTokenCookie` устанавливает cookie `token` во всех контроллерах
    обеспечивая бесшумное обновление.
 
 ### Смешанное хранение CSRF-токена
